+++ src/Modules/Basket/BasketControl.php
@@ -52,9 +52,9 @@ class BasketControl extends Control
 		if ($this->session->may()) {
 			if ($basket['fs_id'] == $this->session->id()) {
 				$requests = $this->basketGateway->listRequests($basket['id'], $this->session->id());
-				$wallPosts = $this->wallposts('basket', $basket['id'], true);
+				$wallPosts = $this->wallposts('basket', $basket['id']);
 			} else {
-				$wallPosts = $this->wallposts('basket', $basket['id'], false);
+				$wallPosts = $this->wallposts('basket', $basket['id']);
 			}
 		}
 		if ($basket['until_ts'] >= time() && $basket['status'] == Status::REQUESTED_MESSAGE_READ) {