+++ CHANGELOG.md
@@ -7,7 +7,7 @@
 - updated bootstrap-vue to v2.0.0-rc20
 - Added Rest endpoint for the current user's profile !880 @dthulke
 - InfluxDB Metrics via UDP !882 @alangecker
-- Only allow basket comments for basket provides !890 @jofranz
+- Only allow basket comments for basket providers !890 @jofranz
 
 ## Bugfixes
 - Fixed a bug in MessageModel.php which caused that conversation members were sometimes not returned !878 @dthulke